pyproject.toml
[project]
name = "userbot-scrapping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "loguru>=0.7.3",
    "pysocks>=1.7.1",
    "python-dotenv>=1.1.0",
    "python-socks[asyncio]>=2.7.1",
    "pytz>=2025.2",
    "telethon>=1.39.0",
    "tgcrypto>=1.2.5",
]

[tool.uv]
constraint-dependencies = ["pyqt5-qt5 <=5.15.2"]
